chore: rename MonthNavigator and dateUtils files to follow naming conventions
